Reset unset tinyint values to default on build
When resetting a tinyint value to its default (often 0), it is no longer added to the YAML file. This resulted in the field being ignored on other installs when running a build, and thus not receiving the updated value.

Now, all tinyint fields that are NOT explicitly set in the YAML config will be reset to the default value.

This ensures that things like enabling/disabling plugins or static templates will be properly applied on remote installations.

This fixes #411.